Don't call boot.sh from run.sh
Cloudfoundry have stopped using boot.sh in their static build pack. We
were calling it from run.sh, which was invoked as a custom start command
from the manifest. This broke things.

I've added the new default start command for the static build pack to
the end of the run.sh script.