Replace route service with a standalone nginx app
Instead of attaching the nginx app as a route service we deploy it
as a separate standalone PaaS application with a cdn-route service
attached to support our environment domains.

This replaces the EC2 AMI creation flow from digitalmarketplace-aws
with a docker container build that follows similar steps.

Docker builds a container with nginx, awslogs, supervisord and python
requirements installed and configuration templates copied to the app
folder.

When the container is launched by CloudFoundry supervisor runs nginx
start up script, which will render the templates using the environment
variables as the context. Similar to the AMI approach, this allows us to
avoid hard-coding any environment configuration into the Docker image,
so that we can use the same public container for all environments.

Dockerfile, awslogs and supervisor configs are relatively similar to the
base docker image, but adding any conditional processing to them
complicates the set up and it could be beneficial to be able to modify
the nginx setup separately from the application ones.