Fix nginx allow/deny block by setting the correct client ip
In AWS our nginx instances were reading the client IP from the
X-Forwarded-For header set by the ELB, which means we only needed
to skip the AWS ELB IP range.

When deployed to PaaS with a CloudFront service the requests travels
through a few proxy servers before getting to nginx, so we need to
include each one of those to the list of trusted IPs.

The request travels through:
* CloudFront, which has a number of IP ranges that are published by AWS.
  We use a script to parse the JSON response and add them to nginx
  configuration when the PaaS instance is launched (to make sure the
  list is up-to-date).
* Some local server, presumably on the PaaS router VM, which ends up
  adding 127.0.0.1 to X-Forwarded-For.
* PaaS router, which uses an internal network range of 10.0.*.*. and
  is the client IP nginx sees in the TCP connection.

We also configure nginx to use the first non-trusted IP from the list
in X-Forwarded-For by switching real_ip_recursive. Otherwise, nginx
will replace the IP address from the TCP connection with the last
record in X-Forwarded-For without checking if it's a trusted proxy.

When starting nginx we render the nginx.conf.j2 template to add the
current list of CloudFront IP ranges. This creates quite a few
set_real_ip_from records, but seems to be the only way to get the
correct element from X-Forwarded-For with nginx.

This approach should give us the correct IP for the external connection
and shouldn't interfere with our IP access restrictions as long as none
of our whitelisted IP ranges overlap with our proxy ranges (so eg we
can't allow access for users coming from CloudFront IP ranges).